Fetching data from database and deserialization implemented. WORKING!
