+++ app/src/main/java/com/example/quizaplication/model/QuizQuestion.kt
@@ -2,6 +2,6 @@ package com.example.quizaplication.model
 
 data class QuizQuestion (
     val prompt : String = "",
-    val options : List<String> = listOf(),
+    val options : Array<String> = arrayOf(),
     val correct : String = ""
 )
