+++ app/src/main/java/com/example/quizaplication/screens/quiz/MultipleChoice.kt
@@ -1,11 +1,9 @@
 package com.example.quizaplication.screens.quiz
 
-import android.util.Log
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.livedata.observeAsState
@@ -18,22 +16,22 @@ import androidx.navigation.NavController
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.example.quizaplication.model.Quiz
 
 @Composable
 fun MultipleChoice(navController: NavController) {
-    val viewModel: QuizViewModelNew = viewModel()
-    //val quiz by viewModel.quiz.observeAsState(initial = null)
-    //val quizQuestions = quiz?.questions?.values?.toList() ?: emptyList()
-
+    val viewModel: QuizViewModel = viewModel()
     val quizQuestions by viewModel.quizQuestions.observeAsState(initial = emptyList())
 
     LaunchedEffect(Unit){
         viewModel.fetchHistoryData()
     }
+
+
         Column(modifier = Modifier.padding(16.dp)) {
         quizQuestions.forEachIndexed { index, question ->
-            Text(text = "${index + 1}. ${question.correct}")
+            Text(text = "${index + 1}. OPTIONS: ${question.options.joinToString(", ")}")
+            Text(text = "${index + 1}. PROMPT: ${question.prompt}")
+            Text(text = "${index + 1}. CORRECT: ${question.correct}")
             Spacer(modifier = Modifier.height(8.dp))
         }
     }
