--- app/src/main/java/com/example/quizaplication/screens/quiz/QuizViewModelNew.kt
@@ -1,33 +1,0 @@
-package com.example.quizaplication.screens.quiz
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-import com.example.quizaplication.model.QuizQuestion
-import com.google.firebase.firestore.ktx.firestore
-import com.google.firebase.ktx.Firebase
-
-class QuizViewModelNew : ViewModel() {
-    private val _quizQuestions = MutableLiveData<List<QuizQuestion>>()
-    val quizQuestions: LiveData<List<QuizQuestion>> get() = _quizQuestions
-
-    fun fetchHistoryData() {
-        val db = Firebase.firestore
-        val historyRef = db.collection("MultipleChoiceQuiz").document("History")
-        historyRef.get().addOnSuccessListener { document ->
-            val questionData = document["questions"] as? Map<String, Any>
-            val questionList = mutableListOf<QuizQuestion>()
-
-            questionData?.forEach { (_, value) ->
-                val questionMap = value as? Map<String, Any>
-                val prompt = questionMap?.keys?.firstOrNull() ?: ""
-                val options = questionMap?.get(prompt) as? List<String> ?: listOf()
-                val correct = questionMap?.get("correct") as? String ?: ""
-
-                questionList.add(QuizQuestion(prompt, options, correct))
-            }
-            _quizQuestions.postValue(questionList)
-        }
-
-    }
-}
